[App/SlackService] Refacto
