+++ src/external-services/slack/slack.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
-import { SlackBlockBuilderService } from './slack-block-builder.service';
 import { SlackService } from './slack.service';
 
 @Module({
   imports: [],
-  providers: [SlackService, SlackBlockBuilderService],
-  exports: [SlackService, SlackBlockBuilderService],
+  providers: [SlackService],
+  exports: [SlackService],
 })
 export class SlackModule {}
